chore(JobsForm): allow EE augmentation
Allow augmentation of placement form section of jobs form with EE features.
